Renamed dpnp_ndarray_types.py to dpnp_ndarray_type.py.
