+++ numba_dpex/dpnp_ndarray/__init__.py
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from ..core.types.dpnp_ndarray_types import *
+from ..core.types.dpnp_ndarray_type import *
 from ..core.typing import typeof
 from . import dpnp_empty  # noqa: isort
 from . import models
